Upgrade to pytorch 1.8 and python 3.7 and fix corresponding errors (#866)
* Upgrade to pytorch 1.8

* AdamW fix

* Remove trunc_normal

* Remove MixtureSameFamily

* Update doc requirements

* AdamW comment

* Remove unnecessary conversion

* Add torchtext

* Changed to python3.7; add torchtext

* Forwarding alf version number

* Fix rendering error in docker with python3.7

* The deadly triad for gym rendering: gym+pyglet+pillow

* More updates

* Add python3.7-dev in readme; minor carla docker update

* Minor fix

+++ alf/algorithms/entropy_target_algorithm.py
@@ -149,8 +149,15 @@ class EntropyTargetAlgorithm(Algorithm):
 
         self._target_entropy = target_entropy
         self._very_slow_update_rate = very_slow_update_rate
-        self._slow_update_rate = torch.tensor(slow_update_rate)
-        self._fast_update_rate = torch.tensor(fast_update_rate)
+
+        # need to explicitly specify dtype to be the same as `self._update_rate`
+        # as required by the `torch.where` function later. This was not needed
+        # in lower version of pytorch (e.g. 1.4) as it will cast a np.float64
+        # to torch.float32.
+        self._slow_update_rate = torch.tensor(
+            slow_update_rate, dtype=torch.float32)
+        self._fast_update_rate = torch.tensor(
+            fast_update_rate, dtype=torch.float32)
 
     def rollout_step(self, distribution_and_step_type, state=None):
         """Rollout step.
